Services.yaml now contains specific parameter called image_directory which refers to the public/images.
added the add.css to style the add product page and its form.
adjusted navbar.css to style the user icon alongside other navbar elements.

add.html.twig holds the routes to both the add page and the route name to the add functions that is called
when saving uploaded photos from the web as a request and stored that public/images directory (through the image_directory parameter)
 under the name of the Product, whereas the local directory (/images/*) gets saved into the database for manipulation.

header.html.twig configured with a js code that manages the dropdown around the user icon's functionality properly.